Take the small findings of the architecture review
- LabelRewriteTest initializes the instruction set itself instead of relying on a constructor
  that happens to do it.
- The lottery weight divides by a long, so that an absurd selectionSpread cannot overflow it.
- Molecule.toString writes the value through formatValue, and the visualizer's palette names
  its value formats by the constants the file defines.
- LabelList no longer justifies its order with a range search that is gone, and example.conf
  no longer speaks of a label rewriting every newborn gets.
- The proposal's outcome records estimateMemoryBytes and that the assembly specification and
  the README stayed as they were.

+++ src/main/java/org/evochora/runtime/label/HammingLabelMatchingStrategy.java
@@ -299,7 +299,7 @@ public class HammingLabelMatchingStrategy implements ILabelMatchingStrategy {
             int flatIndex = own.flatIndexAt(i);
             int distance = coordinates.distance(callerCoords, flatIndex);
             if (selectionSpread > 0) {
-                long weight = Math.max(1, (long) WEIGHT_PRECISION * selectionSpread / (distance + selectionSpread));
+                long weight = Math.max(1, (long) WEIGHT_PRECISION * selectionSpread / ((long) distance + selectionSpread));
                 totalWeight += weight;
                 if (random.nextLong(totalWeight) < weight) {
                     chosen = flatIndex;
